requirements.txt
certifi==2020.6.20
chardet==3.0.4
dnspython==2.0.0
idna==2.10
oauthlib==3.1.0
pymongo==3.11.2
PySocks==1.7.1
python-decouple==3.3
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
tweepy==3.9.0
urllib3==1.25.11
